requirements.txt
Django==5.0.2
python-decouple
django-unfold
model-bakery

ruff